+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- Update npm packages
+
 ## 0.3.2 (2020/08/11)
 
 - Update npm packages